pyproject.toml
[project]
name = "news-digest"
version = "0.1.0"
description = ""
authors = [
    {name = "ainzzorl",email = "ainzzorl@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "feedparser (>=6.0.11,<7.0.0)",
    "html2text (>=2025.4.15,<2026.0.0)",
    "google-api-python-client (>=2.169.0,<3.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "google-auth-oauthlib (>=1.2.2,<2.0.0)",
    "oauth2client (>=4.1.3,<5.0.0)",
    "asyncpraw (>=7.8.1,<8.0.0)",
    "telethon (==1.40.0)",
    "urllib3 (==1.26.6)",
    "boto3 (>=1.38.7,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = true

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"
